backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
PyMuPDF==1.23.3
psycopg2-binary==2.9.9
supabase-py==2.0.0
openai==0.28.1
pydantic==2.4.2
pypdf2==3.0.1
tqdm==4.66.1
aiofiles==23.2.1
requests==2.31.0